Crate merino_integration_tests

Tests for Merino that work by reading from the external API only.

Since the URL endpoints Merino exposes to the world are its public API, and other systems depend on them, the paths used in tests here are important details, and used to keep compatibility.

This is structured as a separate crate so that it produces a single test binary instead of one test per file like would happen if this were merino/tests/.... This improves compilation and test times.

The primary tool used by tests is merino_test, which creates mock servers, sets up the application for testing, and provides helpers to inspect the state of the app. It then calls the test function that is passed to it, providing the above tools as an argument.

#[actix_rt::test]
async fn lbheartbeat_works() {
    merino_test(
        |_| (),
        |TestingTools { test_client, .. }| async move {
            let response = test_client
                .get("/__lbheartbeat__")
                .send()
                .await
                .expect("failed to execute request");

            assert_eq!(response.status(), StatusCode::OK);
            assert_eq!(response.content_length(), Some(0));
        },
    )
    .await
}

Modules

utils
Utilities to aid in testing.

Structs

TestReqwestClient
A wrapper around a [reqwest::client] that automatically sends requests to the test server.

TestingTools
A set of tools for tests, including mock servers and logging helpers.

Statics

VIADUCT_INIT
A marker to track that the viaduct backend has been initialized.

Functions

merino_test
Run a test with a fully configured Merino server.